update utils organized
